requirements.txt
django-filter==24.2
Django==5.0.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0